Move generics to utils.py
